Allows case conversion when object of MemberExpr
Fixes GH-128 cases were not being converted due to the parent node being
a MemberExpression. We do not want member expression properties to be
converted however object's are fair game.

new_collection[1] // should be converted

foo.new_collection // should not

Also applies the fix to snake_case.

This diff also separates the fail cases so each are included in the
coverage report.

And finally, the traverse function was changed so we can use strict
equality when testing if nodes are the same.